PAYARA-4055: Only decorate JAXRS Client for opentracing-enabled clients
Limits the need for decoration of the  client, which breaks the expectations of jaxrs TCK.

All of the decoration and MP Rest Client support for tracing moved to opentracing
package.

Tracing feature remains toggleable at runtime.
