+++ appserver/payara-appserver-modules/jaxrs-client-tracing/pom.xml
@@ -66,16 +66,6 @@
             <artifactId>opentracing-adapter</artifactId>
             <version>${project.version}</version>
         </dependency>
-        <dependency>
-            <groupId>org.eclipse.microprofile.rest.client</groupId>
-            <artifactId>microprofile-rest-client-api</artifactId>
-            <version>${microprofile-rest-client.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.microprofile.opentracing</groupId>
-            <artifactId>microprofile-opentracing-api</artifactId>
-            <version>${microprofile-opentracing.version}</version>
-        </dependency>
     </dependencies>
     
 </project>
