--- appserver/payara-appserver-modules/jaxrs-client-tracing/src/main/resources/META-INF/services/javax.ws.rs.client.ClientBuilder
@@ -1,1 +1,0 @@
-fish.payara.requesttracing.jaxrs.client.decorators.JaxrsClientBuilderDecorator
