+++ appserver/payara-appserver-modules/jaxrs-client-tracing/src/main/resources/META-INF/services/org.glassfish.jersey.internal.spi.ForcedAutoDiscoverable
@@ -38,4 +38,4 @@
 #    holder.
 #
 
-fish.payara.requesttracing.jaxrs.client.mprest.TracingClientListener
+fish.payara.requesttracing.jaxrs.client.JaxrsClientRequestTracingDiscoverable
