+++ appserver/payara-appserver-modules/microprofile/opentracing/src/main/java/fish/payara/microprofile/opentracing/jaxrs/client/JaxrsClientBuilderDecorator.java
@@ -1,43 +1,43 @@
 /*
- * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
- * 
- *    Copyright (c) [2018-2019] Payara Foundation and/or its affiliates. All rights reserved.
- * 
- *     The contents of this file are subject to the terms of either the GNU
- *     General Public License Version 2 only ("GPL") or the Common Development
- *     and Distribution License("CDDL") (collectively, the "License").  You
- *     may not use this file except in compliance with the License.  You can
- *     obtain a copy of the License at
- *     https://github.com/payara/Payara/blob/master/LICENSE.txt
- *     See the License for the specific
- *     language governing permissions and limitations under the License.
- * 
- *     When distributing the software, include this License Header Notice in each
- *     file and include the License file at glassfish/legal/LICENSE.txt.
- * 
- *     GPL Classpath Exception:
- *     The Payara Foundation designates this particular file as subject to the "Classpath"
- *     exception as provided by the Payara Foundation in the GPL Version 2 section of the License
- *     file that accompanied this code.
- * 
- *     Modifications:
- *     If applicable, add the following below the License Header, with the fields
- *     enclosed by brackets [] replaced by your own identifying information:
- *     "Portions Copyright [year] [name of copyright owner]"
- * 
- *     Contributor(s):
- *     If you wish your version of this file to be governed by only the CDDL or
- *     only the GPL Version 2, indicate your decision by adding "[Contributor]
- *     elects to include this software in this distribution under the [CDDL or GPL
- *     Version 2] license."  If you don't indicate a single choice of license, a
- *     recipient has the option to distribute your version of this file under
- *     either the CDDL, the GPL Version 2 or to extend the choice of license to
- *     its licensees as provided above.  However, if you add GPL Version 2 code
- *     and therefore, elected the GPL Version 2 license, then the option applies
- *     only if the new code is made subject to such option by the copyright
- *     holder.
+ *    DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
+ *
+ *    Copyright (c) [2019] Payara Foundation and/or its affiliates. All rights reserved.
+ *
+ *    The contents of this file are subject to the terms of either the GNU
+ *    General Public License Version 2 only ("GPL") or the Common Development
+ *    and Distribution License("CDDL") (collectively, the "License").  You
+ *    may not use this file except in compliance with the License.  You can
+ *    obtain a copy of the License at
+ *    https://github.com/payara/Payara/blob/master/LICENSE.txt
+ *    See the License for the specific
+ *    language governing permissions and limitations under the License.
+ *
+ *    When distributing the software, include this License Header Notice in each
+ *    file and include the License file at glassfish/legal/LICENSE.txt.
+ *
+ *    GPL Classpath Exception:
+ *    The Payara Foundation designates this particular file as subject to the "Classpath"
+ *    exception as provided by the Payara Foundation in the GPL Version 2 section of the License
+ *    file that accompanied this code.
+ *
+ *    Modifications:
+ *    If applicable, add the following below the License Header, with the fields
+ *    enclosed by brackets [] replaced by your own identifying information:
+ *    "Portions Copyright [year] [name of copyright owner]"
+ *
+ *    Contributor(s):
+ *    If you wish your version of this file to be governed by only the CDDL or
+ *    only the GPL Version 2, indicate your decision by adding "[Contributor]
+ *    elects to include this software in this distribution under the [CDDL or GPL
+ *    Version 2] license."  If you don't indicate a single choice of license, a
+ *    recipient has the option to distribute your version of this file under
+ *    either the CDDL, the GPL Version 2 or to extend the choice of license to
+ *    its licensees as provided above.  However, if you add GPL Version 2 code
+ *    and therefore, elected the GPL Version 2 license, then the option applies
+ *    only if the new code is made subject to such option by the copyright
+ *    holder.
  */
-package fish.payara.requesttracing.jaxrs.client.decorators;
+package fish.payara.microprofile.opentracing.jaxrs.client;
 
 import fish.payara.nucleus.requesttracing.RequestTracingService;
 import fish.payara.opentracing.OpenTracingService;
@@ -65,7 +65,7 @@ import org.glassfish.jersey.client.JerseyClientBuilder;
  * 
  * @author Andrew Pielage <andrew.pielage@payara.fish>
  */
-public class JaxrsClientBuilderDecorator extends ClientBuilder {
+class JaxrsClientBuilderDecorator extends ClientBuilder {
 
     public static final String EARLY_BUILDER_INIT = "fish.payara.requesttracing.jaxrs.client.decorators.EarlyBuilderInit";
 
@@ -73,11 +73,20 @@ public class JaxrsClientBuilderDecorator extends ClientBuilder {
     
     /**
      * Initialises a new JerseyClientBuilder and sets it as the decorated object.
+     * @param clientBuilder
      */
-    public JaxrsClientBuilderDecorator() {
-        this.clientBuilder = new JerseyClientBuilder();
+    private JaxrsClientBuilderDecorator(ClientBuilder clientBuilder) {
+        this.clientBuilder = clientBuilder;
     }
-    
+
+    public static ClientBuilder wrap(ClientBuilder clientBuilder) {
+        if (clientBuilder instanceof JaxrsClientBuilderDecorator) {
+            return clientBuilder;
+        } else {
+            return new JaxrsClientBuilderDecorator(clientBuilder);
+        }
+    }
+
     @Override
     public ClientBuilder withConfig(Configuration config) {
         clientBuilder = clientBuilder.withConfig(config);
@@ -138,9 +147,6 @@ public class JaxrsClientBuilderDecorator extends ClientBuilder {
             return clientBuilder.build();
         }
 
-        // Register the Request Tracing filter
-        this.register(JaxrsClientRequestTracingFilter.class);
-
         // Build and return a decorated client
         Client client = this.clientBuilder.build();
 
@@ -162,8 +168,7 @@ public class JaxrsClientBuilderDecorator extends ClientBuilder {
                 ServiceHandle<RequestTracingService> requestTracingHandle = serviceLocator.getServiceHandle(RequestTracingService.class);
                 ServiceHandle<OpenTracingService> openTracingHandle = serviceLocator.getServiceHandle(OpenTracingService.class);
                 return requestTracingHandle != null && openTracingHandle != null
-                        && requestTracingHandle.isActive() && openTracingHandle.isActive()
-                        && requestTracingHandle.getService().isRequestTracingEnabled();
+                        && requestTracingHandle.isActive() && openTracingHandle.isActive();
             }
         } catch (Exception e) {
             // means that we likely cannot do request tracing anyway
