+++ appserver/payara-appserver-modules/microprofile/opentracing/src/main/resources/META-INF/services/org.eclipse.microprofile.opentracing.ClientTracingRegistrarProvider
@@ -1,1 +1,1 @@
-fish.payara.microprofile.opentracing.jaxrs.JaxrsClientRegistrarProvider
+fish.payara.microprofile.opentracing.jaxrs.client.JaxrsClientRegistrarProvider
